video: don't use mmap'ed fb memory as scratch memory
mmap'd fbdev is kinda slow memory as it will end up on the screen (duh).
For scaling, a line was written and then replicated. This is very slow
memcpy'ing from the mmap'd region to itself, and we end up being much
faster by writing the line in a dedicated buffer. Generate the line in
regular memory, then memcpy it multiple times for vertical scaling.

With e.g. an autoscaling factor of 10, I_FinishUpdate goes down from
125ms to 4ms on average. This is less dramatic for a scaling of 4, where
it goes down from ~17ms to 1.5ms (still a 10x improvement).

At any rate, this becomes the difference between hitting 60fps or not.